Task running. Need to change on state and fix stim rampdown
